test: add create tokens flow
