+++ package.json
@@ -38,7 +38,7 @@
   "dependencies": {
     "@radixdlt/babylon-gateway-api-sdk": "^1.0.0-rc.1.4",
     "@radixdlt/connect-button": "0.13.0-alpha.1",
-    "@radixdlt/wallet-sdk": "0.9.0-alpha.1",
+    "@radixdlt/wallet-sdk": "0.9.0-alpha.2",
     "lodash.isequal": "^4.5.0",
     "neverthrow": "^6.0.0",
     "rxjs": "^7.8.0"
